backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
supabase>=2.13.0
slowapi>=0.1.9
python-multipart>=0.0.18
prometheus-fastapi-instrumentator>=7.0.0
pydantic-settings>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.25.0
python-jose[cryptography]>=3.3.0
